Resolve compiler blockers, duplicate modules, and unify strategic blueprints
- Cleanly resolved raw Git merge conflict markers across `FUTURE-DEVELOPMENT-ROADMAP.md` and `README.md`.
- Fixed trailing/unclosed delimiters in `mac.rs`, `secrets.rs`, and `pki.rs`.
- Fixed duplicate `pub mod` declarations, duplicate enum definitions, and type-mismatch transmutation errors (E0512) across package, AI, and driver modules.
- Created `SIGMAOS_AUTONOMOUS_AI_ENGINEERING_SPECIFICATION.md` detailing 18 AI workflows, universal device adapters, and a distro-crushing framework.
- Synchronized all documentation updates across local `WIKI/` and `wiki_repo/` targets.

+++ src/ml/inference.rs
@@ -53,7 +53,7 @@ impl SimpleMLModel {
 
 impl MLModel for SimpleMLModel {
     fn id(&self) -> ModelID { self.id }
-    fn model_type(&self) -> ModelType { unsafe { core::mem::transmute(self.model_type.load(Ordering::SeqCst) as u32) } }
+    fn model_type(&self) -> ModelType { unsafe { core::mem::transmute(self.model_type.load(Ordering::SeqCst)) } }
     fn input_size(&self) -> usize { self.input_size.load(Ordering::SeqCst) }
     fn output_size(&self) -> usize { self.output_size.load(Ordering::SeqCst) }
